Fix : singup page frontend
