+++ frontend-admin/.env.development
@@ -2,4 +2,4 @@
 ENV = 'development'
 
 # base api
-VUE_APP_BASE_API = '/dev-api'
+VUE_APP_BASE_API = 'https://wdc-backend-live.dev.vmv.re/api'
